requirements.txt
future==0.18.3
numpy==1.21.1
pygame==1.9.6
Pillow==9.4.0
poetry==1.3.2
tensorboard
tqdm
torch
torchvision